Rename render.js to view.js
